DEV - Implemented Weapon
Unfortunately the weapon doesn't MOVE in the player's hand, this will have to be solved by having the weapon sprite live on the player GameObject and creating melee attack animations from there and all that.
For now, this system works and is good for the most basic demos (due tonight) so we're good.

:camel: