DEV - Ugly-Fixed AmbushZone's LevelBound-setting
Originally, the values I were getting didn't pull properly from the MAIN CAMERA and not the screen size. I was able to use some match to determine that the desired position was correlated to the 16:9 screen ratio multiplied by the main camera's Size field, I have yet to see how exactly what value lies in-between those variables that is the base that I need to scale up from. For now though, it's "fixed" - it works and works consistently.

:llama: :movie_camera: