DEV - Implemented basic FirePowerup
The ability responds to LAlt for Player 1 - this can be remapped. In any case, my initial work seems to be successful with this one, it's just unable to be visualized without the gizmos on at the moment. I'll have to see what I can do about this besides adding a quick sprite or something.
:llama: