DEV - Implemented a Combustable object
I chose to wrangle with more of Corgi's assets so I made an object that has a Health and resists damage to anything except Fire damage, to which it dies for. This can be changed for animation and things later but for now this is sufficient.
There's still the annoying movement error with the powerup I need to fix, I'm really not sure how to get that right YET.

:llama: :fire: